+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 ## [unreleased]
 - fix: No longer panic at the end of a program when `Context` is dropped
+- refactor: optimize `Graphics::pixels_mut`
+- feat: Add `Graphics::pixels` method and `Pixels` type